+++ frontend/src/pages/NotFound.jsx
@@ -1,0 +1,10 @@
+function NotFound() {
+  return (
+    <main className="w-full flex justify-center items-center h-[67vh] flex-col">
+      <img className="w-16" src="/images/borboleta.png" alt="Imagem de erro" />
+      <h1 className="font-viminalis">Página não encontrada!</h1>
+    </main>
+  )
+}
+
+export default NotFound